requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
google-auth==2.37.0
google-genai==0.3.0
idna==3.10
lxml==5.3.0
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
requests==2.32.3
rsa==4.9
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.1
XlsxWriter==3.2.0
